+++ src/main/java/CalculatorApplication.java
@@ -1,12 +1,15 @@
 import domain.Calculator;
+import domain.Expression;
 import view.InputView;
 import view.OutputView;
 
 public class CalculatorApplication {
     public static void main(String[] args) {
-        Calculator calculator = InputView.inputExpressions();
-        int result = calculator.calculate(calculator.getNumbers(), calculator.getOperators());
+        Calculator calculator = new Calculator();
+        Expression expression = InputView.inputExpressions();
 
+        int result = calculator.calculate(expression.getNumbers(), expression.getOperators());
         OutputView.printResult(result);
     }
 }
+
