+++ src/main/java/view/InputView.java
@@ -1,14 +1,13 @@
 package view;
 
-import domain.Calculator;
-
+import domain.Expression;
 import java.util.Scanner;
 
 public class InputView {
-    public static Calculator inputExpressions() {
+    public static Expression inputExpressions() {
         try (Scanner scanner = new Scanner(System.in)) {
             System.out.println("계산식을 입력해주세요.");
-            return new Calculator(scanner.nextLine());
+            return new Expression(scanner.nextLine());
         }
     }
 }
